Consolidate all project config files
