+++ clients/admin-ui/.env.development
@@ -1,1 +1,3 @@
 NEXT_PUBLIC_FIDESOPS_API=http://0.0.0.0:8080
+NEXT_PUBLIC_FIDESCTL_API=/api/v1
+NEXT_PUBLIC_FIDESCTL_API_SERVER=http://0.0.0.0:8080
