+++ clients/admin-ui/.eslintignore
@@ -1,5 +1,5 @@
 node_modules
 dist
-.eslintrc.js
+.eslintrc.json
 next.config.js
 jest.config.js
