+++ clients/admin-ui/tsconfig.json
@@ -1,24 +1,43 @@
 {
   "compilerOptions": {
     "allowJs": true,
-    "baseUrl": "../ops/admin-ui",
+    "baseUrl": ".",
     "downlevelIteration": true,
     "esModuleInterop": true,
     "forceConsistentCasingInFileNames": true,
     "incremental": true,
     "isolatedModules": true,
     "jsx": "preserve",
-    "lib": ["dom", "dom.iterable", "esnext"],
+    "lib": [
+      "dom",
+      "dom.iterable",
+      "esnext"
+    ],
     "module": "esnext",
     "moduleResolution": "node",
     "noEmit": true,
     "paths": {
-      "common/*": ["src/features/common/*"],
-      "connection-type/*": ["src/features/connection-type/*"],
-      "datastore-connections/*": ["src/features/datastore-connections/*"],
-      "privacy-requests/*": ["src/features/privacy-requests/*"],
-      "subject-request/*": ["src/features/subject-request/*"],
-      "user-management/*": ["src/features/user-management/*"]
+      "~/*": [
+        "src/*"
+      ],
+      "common/*": [
+        "src/features/common/*"
+      ],
+      "connection-type/*": [
+        "src/features/connection-type/*"
+      ],
+      "datastore-connections/*": [
+        "src/features/datastore-connections/*"
+      ],
+      "privacy-requests/*": [
+        "src/features/privacy-requests/*"
+      ],
+      "subject-request/*": [
+        "src/features/subject-request/*"
+      ],
+      "user-management/*": [
+        "src/features/user-management/*"
+      ]
     },
     "resolveJsonModule": true,
     "skipLibCheck": true,
@@ -27,6 +46,14 @@
     "target": "es5"
   },
   "include": [
-    "next-env.d.ts", "**/*.ts", "**/*.tsx", "**/*.d.ts"],
-  "exclude": ["node_modules"]
+    "next-env.d.ts",
+    "**/*.ts",
+    "**/*.tsx",
+    "**/*.d.ts"
+  ],
+  "exclude": [
+    "node_modules",
+    "cypress/**/*.ts",
+    "../../admin-ui/cypress.config.ts"
+  ]
 }
