--- clients/ctl/admin-ui/.env.production
@@ -1,2 +1,0 @@
-NEXT_PUBLIC_FIDESCTL_API=/api/v1
-NEXT_PUBLIC_FIDESCTL_API_SERVER=http://0.0.0.0:8080
