--- clients/ctl/admin-ui/.prettierrc.json
@@ -1,3 +1,0 @@
-{
-  "tabWidth": 2
-}
